+++ data-transfer/src/main/java/it/eng/datatransfer/service/AgreementService.java
@@ -3,42 +3,57 @@ package it.eng.datatransfer.service;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import it.eng.datatransfer.exceptions.AgreementNotFoundException;
+import it.eng.datatransfer.model.TransferProcess;
 import it.eng.datatransfer.repository.TransferProcessRepository;
-import it.eng.datatransfer.repository.TransferRequestMessageRepository;
+import it.eng.tools.client.rest.OkHttpRestClient;
+import it.eng.tools.controller.ApiEndpoints;
+import it.eng.tools.response.GenericApiResponse;
+import it.eng.tools.util.CredentialUtils;
 import lombok.extern.slf4j.Slf4j;
 
 @Service
 @Slf4j
 public class AgreementService {
 	
-	Map<String, Boolean> agreementStorage = new HashMap<>();
-	private TransferRequestMessageRepository transferRequestMessageRepository;
-	private TransferProcessRepository transferProcessRepository;
+	@Value("${server.port}")
+	private String serverPort;
 	
-	public AgreementService(TransferRequestMessageRepository transferRequestMessageRepository,
-			TransferProcessRepository transferProcessRepository) {
+	Map<String, Boolean> agreementStorage = new HashMap<>();
+	private TransferProcessRepository transferProcessRepository;
+	private OkHttpRestClient okHttpRestClient;
+	private CredentialUtils credentialUtils;
+	
+	public AgreementService(TransferProcessRepository transferProcessRepository, OkHttpRestClient okHttpRestClient,
+			CredentialUtils credentialUtils) {
 		super();
-		this.transferRequestMessageRepository = transferRequestMessageRepository;
 		this.transferProcessRepository = transferProcessRepository;
+		this.okHttpRestClient = okHttpRestClient;
+		this.credentialUtils = credentialUtils;
 	}
 
 	public boolean isAgreementValid(String consumerPid, String providerPid) {
-		
-//		String agreementId = transferRequestMessageRepository.findByConsumerPid(consumerPid)
-//				.map(trm -> trm.getAgreementId())
-//				.orElseThrow(() -> new AgreementNotFoundException("Agreement for cosnumerPid '"+ consumerPid +
-//						"' and providerPid '" + providerPid + "' not found"));
-		String agreementId = transferProcessRepository.findByConsumerPidAndProviderPid(consumerPid, providerPid)
-				.map(trm -> trm.getAgreementId())
+		TransferProcess transferProcess = transferProcessRepository.findByConsumerPidAndProviderPid(consumerPid, providerPid)
 				.orElseThrow(() -> new AgreementNotFoundException("Agreement for cosnumerPid '"+ consumerPid +
 						"' and providerPid '" + providerPid + "' not found", consumerPid, providerPid));
 		
+		String agreementId = transferProcess.getAgreementId();
 		// TODO once usage policy enforcement is done, call should be made here
-		boolean isValid = checkIfAgreementIsValid(agreementId);
-		return isValid;
+
+		GenericApiResponse<String> response = okHttpRestClient.sendRequestProtocol("http://localhost:" + serverPort 
+			+ ApiEndpoints.NEGOTIATION_AGREEMENTS_V1 + "/" + agreementId + "/valid", 
+				null, 
+				credentialUtils.getAPICredentials());
+        
+		if (!response.isSuccess()) {
+			log.info("Agreement is not valid");
+			return false;
+		}
+		
+		return true;
 //		if(!agreementStorage.containsKey(agreementId)) {
 //			agreementStorage.put(agreementId, isVaid);
 //		}